House debugger in its own file
