+++ skema/text_reading/mention_linking/gromet_linker/comment_aligner.py
@@ -4,6 +4,7 @@ from automates.program_analysis.JSON2GroMEt.json2gromet import json_to_gromet
 from collections import defaultdict
 from typing import Optional
 
+from .comment_debugger import CommentDebugger
 from .debug_info import DebugInfo
 from .gromet_helper import GrometHelper
 from .source_comments import SourceComments
@@ -12,37 +13,6 @@ from .utils import get_code_file_ref, build_comment_metadata, build_tr_mention_m
 from .variable_name_matcher import VariableNameMatcher
 
 class CommentAligner():
-
-	class Debugger():
-		def add_info(self, info):
-			pass
-
-		def debug():
-			pass
-	
-	class FakeDebugger(Debugger):
-		pass
-
-	class RealDebugger(Debugger):
-		def __init__(self):
-			self.debug_info = list()
-
-		def add_info(self, info):
-			if info:
-				self.debug_info.append(info)
-
-		def debug(self):
-			# Aggregate the debug info per name.
-			grouped_info = defaultdict(list)
-			for info in self.debug_info:
-				key = info.name if info.name else ''
-				grouped_info[key].append(info)
-
-			# Print them in lexicographical order.
-			for key in sorted(grouped_info):
-				for info in grouped_info[key]:
-					info.print()
-
 
 	def __init__(self, time_stamper, uid_stamper, gromet_path: str, comments_path: str, extractions_path: str, embeddings_path: str):
 		self.time_stamper = time_stamper
@@ -55,12 +25,11 @@ class CommentAligner():
 
 	# Return new gromet_fn_module that has been aligned and linked.
 	def align(self, debug: bool = False):
-		debugger = self.RealDebugger() if debug else self.FakeDebugger()
+		debugger = CommentDebugger.create(debug)
 		# This is read anew here, because it will be modified before return.
 		gromet_fn_module = json_to_gromet(self.gromet_path)
 
 		container_gromet_box_function = None
-		gromet_fn = None
 
 		def get_info(gromet_object, key):
 			condition = container_gromet_box_function and container_gromet_box_function.function_type == "FUNCTION"
@@ -71,6 +40,7 @@ class CommentAligner():
 
 			line_range = GrometHelper.get_element_line_numbers(gromet_object, gromet_fn_module)
 
+			# Get both kinds?
 			aligned_comments = self.enhance_attribute_with_comments(gromet_object, key, container_gromet_box_function, gromet_fn_module, line_range)
 			info = self.step2(gromet_object, gromet_fn_module, line_range, aligned_docstrings, aligned_comments)
 			debugger.add_info(info)
@@ -128,7 +98,7 @@ class CommentAligner():
 		return comments
 
 	# This gromet_object should be either a gromet_box_function or gromet_port.
-	def enhance_attribute_with_comments(self, gromet_object, key, container_gromet_box_function, gromet_fn_module, line_range: range) -> Optional[DebugInfo]:
+	def enhance_attribute_with_comments(self, gromet_object, key, container_gromet_box_function, gromet_fn_module, line_range: range):
 		# Identify the variables with the same name to each output port.
 		# In case of a tie, resolve the correct variable using the containing line spans.
 
@@ -165,7 +135,7 @@ class CommentAligner():
 		for d in aligned_docstrings:
 			build_comment_metadata(self.time_stamper, d, code_file_ref, gromet_object, gromet_fn_module)
 		# linked text reading mentions
-		for m in aligned_mentions:
+		for m in 	aligned_mentions:
 			doc_file_ref = get_doc_file_ref(self.time_stamper, self.uid_stamper, m, self.linker, gromet_fn_module)
 			build_tr_mention_metadata(self.time_stamper, m, doc_file_ref, gromet_object, gromet_fn_module)
